Move mission-specific functions to new module
