+++ stingray/io.py
@@ -16,8 +16,16 @@ from astropy.io import fits as pf
 import stingray.utils as utils
 from stingray.loggingconfig import setup_logger
 
-from .utils import assign_value_if_none, is_string, order_list_of_arrays, is_sorted
+from .utils import (
+    assign_value_if_none,
+    is_string,
+    order_list_of_arrays,
+    is_sorted,
+    make_dictionary_lowercase,
+)
 from .gti import get_gti_from_all_extensions, load_gtis
+
+from .mission_support import read_mission_info, rough_calibration
 
 # Python 3
 import pickle
@@ -93,50 +101,6 @@ def pi_to_energy(pis, rmf_file):
     return es
 
 
-def rough_calibration(pis, mission):
-    """Make a rough conversion between PI channel and energy.
-
-    Only works for NICER, NuSTAR, IXPE, and XMM.
-
-    Parameters
-    ----------
-    pis: float or array of floats
-        PI channels in data
-    mission: str
-        Mission name
-
-    Returns
-    -------
-    energies : float or array of floats
-        Energy values
-
-    Examples
-    --------
-    >>> rough_calibration(0, 'nustar')
-    1.6
-    >>> rough_calibration(0.0, 'ixpe')
-    0.0
-    >>> # It's case-insensitive
-    >>> rough_calibration(1200, 'XMm')
-    1.2
-    >>> rough_calibration(10, 'asDf')
-    Traceback (most recent call last):
-        ...
-    ValueError: Mission asdf not recognized
-    >>> rough_calibration(100, 'nicer')
-    1.0
-    """
-    if mission.lower() == "nustar":
-        return pis * 0.04 + 1.6
-    elif mission.lower() == "xmm":
-        return pis * 0.001
-    elif mission.lower() == "nicer":
-        return pis * 0.01
-    elif mission.lower() == "ixpe":
-        return pis / 375 * 15
-    raise ValueError(f"Mission {mission.lower()} not recognized")
-
-
 def get_file_extension(fname):
     """Get the extension from the file name.
 
@@ -196,75 +160,6 @@ def high_precision_keyword_read(hdr, keyword):
         return value
     except KeyError:
         return None
-
-
-def _patch_mission_info(info, mission=None):
-    """Add some information that is surely missing in xselect.mdb.
-
-    Examples
-    --------
-    >>> info = {'gti': 'STDGTI'}
-    >>> new_info = _patch_mission_info(info, mission=None)
-    >>> assert new_info['gti'] == info['gti']
-    >>> new_info = _patch_mission_info(info, mission="xmm")
-    >>> new_info['gti']
-    'STDGTI,GTI0'
-    >>> new_info = _patch_mission_info(info, mission="xte")
-    >>> new_info['ecol']
-    'PHA'
-    """
-    if mission is None:
-        return info
-    if mission.lower() == "xmm" and "gti" in info:
-        info["gti"] += ",GTI0"
-    if mission.lower() == "xte" and "ecol" in info:
-        info["ecol"] = "PHA"
-        info["ccol"] = "PCUID"
-    return info
-
-
-def read_mission_info(mission=None):
-    """Search the relevant information about a mission in xselect.mdb."""
-    curdir = os.path.abspath(os.path.dirname(__file__))
-    fname = os.path.join(curdir, "datasets", "xselect.mdb")
-
-    # If HEADAS is defined, search for the most up-to-date version of the
-    # mission database
-    if os.getenv("HEADAS"):
-        hea_fname = os.path.join(os.getenv("HEADAS"), "bin", "xselect.mdb")
-        if os.path.exists(hea_fname):
-            fname = hea_fname
-    if mission is not None:
-        mission = mission.lower()
-
-    db = {}
-    with open(fname) as fobj:
-        for line in fobj.readlines():
-            line = line.strip()
-            if mission is not None and not line.lower().startswith(mission):
-                continue
-            if line.startswith("!") or line == "":
-                continue
-            allvals = line.split()
-            string = allvals[0]
-            value = allvals[1:]
-            if len(value) == 1:
-                value = value[0]
-
-            data = string.split(":")[:]
-            if mission is None:
-                if data[0] not in db:
-                    db[data[0]] = {}
-                previous_db_step = db[data[0]]
-            else:
-                previous_db_step = db
-            data = data[1:]
-            for key in data[:-1]:
-                if key not in previous_db_step:
-                    previous_db_step[key] = {}
-                previous_db_step = previous_db_step[key]
-            previous_db_step[data[-1]] = value
-    return _patch_mission_info(db, mission)
 
 
 def _case_insensitive_search_in_list(string, list_of_strings):
@@ -363,9 +258,8 @@ def get_key_from_mission_info(info, key, default, inst=None, mode=None):
     >>> get_key_from_mission_info(info, "ghghg", "BU", inst="C", mode="M1")
     'BU'
     """
-    filt_info = copy.deepcopy(info)
+    filt_info = make_dictionary_lowercase(info, recursive=True)
     key = key.lower()
-    filt_info = dict([(k.lower(), v) for k, v in filt_info.items()])
     if inst is not None:
         inst = inst.lower()
         if inst in filt_info:
@@ -374,7 +268,7 @@ def get_key_from_mission_info(info, key, default, inst=None, mode=None):
     if mode is not None:
         mode = mode.lower()
         if mode in filt_info:
-            filt_info.update(filt_info[inst][mode])
+            filt_info.update(filt_info[mode])
             filt_info.pop(mode)
 
     if key in filt_info:
