+++ stingray/mission_support/__init__.py
@@ -1,0 +1,1 @@
+from .missions import rough_calibration, read_mission_info
